+++ src/FunFair.CodeAnalysis/ReThrowingExceptionShouldSpecifyInnerExceptionDiagnosticsAnalyzer.cs
@@ -1,4 +1,7 @@
-﻿using System.Collections.Immutable;
+﻿using System.Collections.Generic;
+using System.Collections.Immutable;
+using System.Linq;
+using FunFair.CodeAnalysis.Helpers;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -12,12 +15,12 @@ namespace FunFair.CodeAnalysis
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
     public sealed class ReThrowingExceptionShouldSpecifyInnerExceptionDiagnosticsAnalyzer : DiagnosticAnalyzer
     {
-        private const string CATEGORY = "Classes";
+        private const string CATEGORY = "Exceptions";
 
-        private static readonly DiagnosticDescriptor Rule = RuleHelpers.CreateRule(code: Rules.RuleClassesShouldBeStaticSealedOrAbstract,
+        private static readonly DiagnosticDescriptor Rule = RuleHelpers.CreateRule(code: Rules.RuleMustPassInterExceptionToExceptionsThrownInCatchBlock,
                                                                                    category: CATEGORY,
-                                                                                   title: "Classes should be static, sealed or abstract",
-                                                                                   message: "Classes should be static, sealed or abstract");
+                                                                                   title: "Pass an a inner exception when thrown from a catch clause",
+                                                                                   message: "Provide '{0}' as a inner exception when throw from the catch clauses");
 
         /// <inheritdoc />
         public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => new[] {Rule}.ToImmutableArray();
@@ -33,21 +36,82 @@ namespace FunFair.CodeAnalysis
 
         private static void PerformCheck(CompilationStartAnalysisContext compilationStartContext)
         {
-            compilationStartContext.RegisterSyntaxNodeAction(action: MustBeReadOnly, SyntaxKind.ClassDeclaration);
+            compilationStartContext.RegisterSyntaxNodeAction(action: MustBeReadOnly, SyntaxKind.CatchClause);
         }
 
         private static void MustBeReadOnly(SyntaxNodeAnalysisContext syntaxNodeAnalysisContext)
         {
-            if (!(syntaxNodeAnalysisContext.Node is ClassDeclarationSyntax classDeclarationSyntax))
+            if (syntaxNodeAnalysisContext.Node is CatchClauseSyntax catchClause)
             {
+                string? exceptionVariable = catchClause.Declaration?.Identifier.Text;
+
+                if (exceptionVariable == null)
+                {
+                    return;
+                }
+
+                IReadOnlyList<ExpressionSyntax> allExpressions = GetAllThrowExpressions(catchClause.Block);
+
+                if (allExpressions.Count == 0)
+                {
+                    return;
+                }
+
+                foreach (ExpressionSyntax expression in allExpressions)
+                {
+                    if (expression is ObjectCreationExpressionSyntax objectCreationExpression)
+                    {
+                        TryToReportDiagnostic(argumentListSyntax: objectCreationExpression.ArgumentList,
+                                              exceptionVariable: exceptionVariable,
+                                              syntaxNodeContext: syntaxNodeAnalysisContext,
+                                              objectCreationExpression: objectCreationExpression);
+                    }
+                    else if (expression is InvocationExpressionSyntax invocationExpressionSyntax)
+                    {
+                        TryToReportDiagnostic(argumentListSyntax: invocationExpressionSyntax.ArgumentList,
+                                              exceptionVariable: exceptionVariable,
+                                              syntaxNodeContext: syntaxNodeAnalysisContext,
+                                              objectCreationExpression: invocationExpressionSyntax);
+                    }
+                }
+            }
+        }
+
+        private static void TryToReportDiagnostic(ArgumentListSyntax? argumentListSyntax,
+                                                  string exceptionVariable,
+                                                  SyntaxNodeAnalysisContext syntaxNodeContext,
+                                                  ExpressionSyntax objectCreationExpression)
+        {
+            if (argumentListSyntax == null || argumentListSyntax.Arguments.Count == 0)
+            {
+                ReportDiagnostic(exceptionVariable: exceptionVariable, syntaxNodeContext: syntaxNodeContext, objectCreationExpression.GetLocation());
+
                 return;
             }
 
-            // if (!classDeclarationSyntax.Modifiers.Any(IsWhiteListedClassModifier))
-            // {
-            syntaxNodeAnalysisContext.ReportDiagnostic(Diagnostic.Create(descriptor: Rule, classDeclarationSyntax.GetLocation()));
+            if (argumentListSyntax.Arguments.Any(x => x.Expression is IdentifierNameSyntax identifier && identifier.Identifier.Text == exceptionVariable))
+            {
+                ReportDiagnostic(exceptionVariable: exceptionVariable, syntaxNodeContext: syntaxNodeContext, objectCreationExpression.GetLocation());
+            }
+        }
 
-            // }
+        private static void ReportDiagnostic(string exceptionVariable, SyntaxNodeAnalysisContext syntaxNodeContext, Location location)
+        {
+            syntaxNodeContext.ReportDiagnostic(Diagnostic.Create(descriptor: Rule, location: location, exceptionVariable));
+        }
+
+        private static IReadOnlyList<ExpressionSyntax> GetAllThrowExpressions(BlockSyntax codeBlock)
+        {
+            IEnumerable<ExpressionSyntax> expressionFromThrowStatements = codeBlock.DescendantNodes()
+                                                                                   .OfType<ThrowStatementSyntax>()
+                                                                                   .Select(x => x.Expression)
+                                                                                   .RemoveNulls();
+            IEnumerable<ExpressionSyntax> expressionFromThrowExpressions = codeBlock.DescendantNodes()
+                                                                                    .OfType<ThrowExpressionSyntax>()
+                                                                                    .Select(x => x.Expression);
+
+            return expressionFromThrowStatements.Concat(expressionFromThrowExpressions)
+                                                .ToArray();
         }
     }
 }